+++ src/device/wobulb.ts
@@ -4,12 +4,65 @@
  */
 import { Buffer } from 'buffer';
 
-import { SwitchbotDevice } from '../switchbot.js';
+import { SwitchbotDevice } from '../device.js';
 
 /**
  * @see https://github.com/OpenWonderLabs/SwitchBotAPI-BLE/blob/latest/devicetypes/colorbulb.md
  */
 export class WoBulb extends SwitchbotDevice {
+  static parseServiceData(manufacturerData, onlog) {
+    if (manufacturerData.length !== 13) {
+      if (onlog && typeof onlog === 'function') {
+        onlog(
+          `[parseServiceDataForWoBulb] Buffer length ${manufacturerData.length} !== 13!`,
+        );
+      }
+      return null;
+    }
+    const byte1 = manufacturerData.readUInt8(1);//power and light status
+    //const byte2 = manufacturerData.readUInt8(2);//bulb brightness
+    const byte3 = manufacturerData.readUInt8(3);//bulb R
+    const byte4 = manufacturerData.readUInt8(4);//bulb G
+    const byte5 = manufacturerData.readUInt8(5);//bulb B
+    const byte6 = manufacturerData.readUInt8(6);//bulb temperature
+    const byte7 = manufacturerData.readUInt8(7);
+    const byte8 = manufacturerData.readUInt8(8);
+    const byte9 = manufacturerData.readUInt8(9);
+    const byte10 = manufacturerData.readUInt8(10);//bulb mode
+
+    const power = byte1;
+    const red = byte3;
+    const green = byte4;
+    const blue = byte5;
+    const color_temperature = byte6;
+    const state = byte7 & 0b01111111 ? true : false;
+    const brightness = byte7 & 0b01111111;
+    const delay = byte8 & 0b10000000;
+    const preset = byte8 & 0b00001000;
+    const color_mode = byte8 & 0b00000111;
+    const speed = byte9 & 0b01111111;
+    const loop_index = byte10 & 0b11111110;
+
+    const data = {
+      model: 'u',
+      modelName: 'WoBulb',
+      color_temperature: color_temperature,
+      power: power,
+      state: state,
+      red: red,
+      green: green,
+      blue: blue,
+      brightness: brightness,
+      delay: delay,
+      preset: preset,
+      color_mode: color_mode,
+      speed: speed,
+      loop_index: loop_index,
+    };
+
+    return data;
+  }
+
   /**
    * @returns {Promise<boolean>} resolves with a boolean that tells whether the plug in ON(true) or OFF(false)
    */
@@ -48,7 +101,7 @@ export class WoBulb extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target brightness percentage is incorrect: ' +
-              typeof brightness,
+            typeof brightness,
           ),
         );
       });
@@ -70,7 +123,7 @@ export class WoBulb extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target color_temperature percentage is incorrect: ' +
-              typeof color_temperature,
+            typeof color_temperature,
           ),
         );
       });
@@ -92,7 +145,7 @@ export class WoBulb extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target brightness percentage is incorrect: ' +
-              typeof brightness,
+            typeof brightness,
           ),
         );
       });
@@ -102,7 +155,7 @@ export class WoBulb extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target red is incorrect: ' +
-              typeof red,
+            typeof red,
           ),
         );
       });
@@ -112,7 +165,7 @@ export class WoBulb extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target green is incorrect: ' +
-              typeof green,
+            typeof green,
           ),
         );
       });
@@ -122,7 +175,7 @@ export class WoBulb extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target blue is incorrect: ' +
-              typeof blue,
+            typeof blue,
           ),
         );
       });
@@ -175,7 +228,7 @@ export class WoBulb extends SwitchbotDevice {
             reject(
               new Error(
                 'Expecting a 2-byte response, got instead: 0x' +
-                  res_buf.toString('hex'),
+                res_buf.toString('hex'),
               ),
             );
           }
